[ethers-v4-v5] move codegen contract factories to subdirectory
