+++ packages/target-ethers-v4/src/codegen/index.ts
@@ -94,7 +94,7 @@ export function codegenContractFactory(contract: Contract, abi: any, bytecode?: 
   ${ethersUtilsImportLine}
 
   import { TransactionOverrides } from ".";
-  import { ${contract.name} } from "./${contract.name}";
+  import { ${contract.name} } from "../${contract.name}";
 
   export class ${contract.name}Factory extends ContractFactory {
     ${generateFactoryConstructor(contract, bytecode)}
@@ -128,7 +128,7 @@ export function codegenAbstractContractFactory(contract: Contract, abi: any): st
   import { Contract, Signer } from "ethers";
   import { Provider } from "ethers/providers";
 
-  import { ${contract.name} } from "./${contract.name}";
+  import { ${contract.name} } from "../${contract.name}";
 
   export class ${contract.name}Factory {
     static connect(address: string, signerOrProvider: Signer | Provider): ${contract.name} {
