+++ packages/target-ethers-v4/src/index.ts
@@ -102,7 +102,7 @@ export default class Ethers extends TsGeneratorPlugin {
 
   genContractFactoryFile(contract: Contract, abi: any, bytecode?: BytecodeWithLinkReferences) {
     return {
-      path: join(this.outDirAbs, `${contract.name}Factory.ts`),
+      path: join(this.outDirAbs, 'factories', `${contract.name}Factory.ts`),
       contents: codegenContractFactory(contract, abi, bytecode),
     }
   }
@@ -113,7 +113,7 @@ export default class Ethers extends TsGeneratorPlugin {
     const abstractFactoryFiles = Object.keys(this.contractCache).map((contractName) => {
       const { contract, abi } = this.contractCache[contractName]
       return {
-        path: join(this.outDirAbs, `${contract.name}Factory.ts`),
+        path: join(this.outDirAbs, 'factories', `${contract.name}Factory.ts`),
         contents: codegenAbstractContractFactory(contract, abi),
       }
     })
